+++ src/utf.c
@@ -465,7 +465,7 @@ char *sqlite3Utf16to8(sqlite3 *db, const void *z, int nByte){
 ** flag set.
 */
 #ifdef SQLITE_ENABLE_STAT2
-char *sqlite3Utf8to16(sqlite3 *db, int enc, char *z, int n, int *pnOut){
+char *sqlite3Utf8to16(sqlite3 *db, u8 enc, char *z, int n, int *pnOut){
   Mem m;
   memset(&m, 0, sizeof(m));
   m.db = db;
